Refactor "apps"
This commit inverts the relationship between the "main" function and
"the app". Now each app (file in the apps directory) is in full contol
of its main function. The shared parts of the two apps has now been
extracted to a "board" module.

The "main_avr" app and the "app" module have now become the "gui" app
instead. The "display_test" app has been renamed to "rapidbw", because
it repeatedly writes the screen completely black and white as quickly as
possible.